src/python/requirements.txt
numpy>=1.20
scipy>=1.7
torch>=1.10
matplotlib>=3.5
PyYAML>=5.4
